fixed user logout
