+++ frontend/src/components/Pages/ProjectPage.js
@@ -1,8 +1,8 @@
 /* Main Component displays either project interface or project selection menu */
+
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import { shallowEqual, useSelector } from "react-redux";
-import useCheckUser from "../../hooks/useCheckUser";
 import "../../styles/ProjectPage.css";
 import Navbar from "../Navbar";
 import PageHeader from "../PageHeader";
@@ -12,10 +12,6 @@ import StageCreator from "../StageCreator";
 export default function ProjectPage() {
   const projectId = useSelector((state) => state.project.id);
   const stages = useSelector((state) => state.project.stages, shallowEqual);
-
-  console.log("project page render");
-
-  useCheckUser();
 
   const stageList = stages
     ? stages.map((stage, index) => {
@@ -29,6 +25,7 @@ export default function ProjectPage() {
         );
       })
     : null;
+
   return (
     <DndProvider backend={HTML5Backend}>
       <PageHeader />
