+++ frontend/src/components/Pages/UserPage.js
@@ -1,8 +1,13 @@
-import useCheckUser from "../../hooks/useCheckUser";
+/* This page should contain user settings control as well as list of projects */
+
+import PageHeader from "../PageHeader";
 import ProjectsList from "../ProjectsList";
 
 export default function UserPage() {
-  useCheckUser();
-
-  return <ProjectsList />;
+  return (
+    <>
+      <PageHeader />
+      <ProjectsList />
+    </>
+  );
 }
