+++ frontend/src/utils/ItemTypes.js
@@ -1,6 +1,4 @@
 export const ItemTypes = {
   PROJECTTAB: "projectTab",
-  TEST: "test",
-
   //TODO: add stage, item
 };
